Visual improvements and general functionalities
• Buttons in views.
• Implemented a new service that searches for clients without an address to be used in AddressCreate and AddressEdit.
• Customers with orders cannot be deleted.
• Deleting the client will delete the associated address.
• The address associated with a customer can only be deleted if the customer has no orders.
• Customers who already have an associated address will not appear on the address creation screen.
• Added edit addresses from Client/Details.
​

+++ Controllers/ClientController.cs
@@ -189,11 +189,23 @@ namespace Examen.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+           
+            var hasOrders = await _clientService.ClientHasOrdersAsync(id);
+            if (hasOrders)
+            {
+                ModelState.AddModelError(string.Empty, "No se puede eliminar clientes con pedidos asociados.");
+                ClientDeleteViewModel clientDelete = await _clientService.GetClientDeleteByIdAsync(id);
+
+                return View(clientDelete);
+            }
+
             var success = await _clientService.DeleteClientAsync(id);
 
             if (!success)
             {
-                return RedirectToAction(nameof(Index));
+                ClientDeleteViewModel clientDelete = await _clientService.GetClientDeleteByIdAsync(id);
+                
+                return View(clientDelete);
             }
 
             return RedirectToAction(nameof(Index));
